pyproject.toml
[tool.poetry]
name = "The Finance Crew"
version = "0.1.0"
description = "Ai Crew for Brad"
authors = ["Emilio <emiliosanalvarez07@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
appdirs = "1.4.4"
attrs = "23.2.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
blinker = "1.7.0"
cachetools = "5.3.3"
certifi = "2024.2.2"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
crewai= "0.11.0"
dataclasses-json = "0.6.4"
decorator = "5.1.1"
Deprecated = "1.2.14"
deprecation = "2.1.0"
distro = "1.9.0"
docstring-parser = "0.15"
emoji = "2.10.1"
faiss-cpu = "1.7.4"
filetype = "1.2.0"
Flask = "3.0.2"
fpdf = "1.7.2"
frozendict = "2.4.0"
frozenlist = "1.4.1"
geojson = "2.5.0"
google-ai-generativelanguage = "0.4.0"
google-api-core = "2.17.1"
google-auth = "2.28.2"
google-generativeai = "0.4.1"
googleapis-common-protos = "1.62.0"
grpcio = "1.62.1"
grpcio-status = "1.62.1"
h11 = "0.14.0"
html5lib = "1.1"
httpcore = "1.0.2"
httpx = "0.26.0"
idna = "3.6"
importlib-metadata = "6.11.0"
instructor = "0.5.2"
itsdangerous = "2.1.2"
Jinja2 = "3.1.3"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
lancedb = "0.6.3"
langchain = "0.1.0"
langchain-community = "0.0.10"
langchain-core = "0.1.8"
langchain-google-genai = "0.0.11"
langchain-openai = "0.0.2.post1"
langdetect = "1.0.9"
langsmith = "0.0.91"
lxml = "5.1.0"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
marshmallow = "3.20.2"
mdurl = "0.1.2"
multidict = "6.0.5"
multitasking = "0.0.11"
mypy-extensions = "1.0.0"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.7.1"
opentelemetry-api = "1.22.0"
opentelemetry-exporter-otlp-proto-common = "1.22.0"
opentelemetry-exporter-otlp-proto-http = "1.22.0"
opentelemetry-proto = "1.22.0"
opentelemetry-sdk = "1.22.0"
opentelemetry-semantic-conventions = "0.43b0"
overrides = "7.7.0"
packaging = "23.2"
pandas = "2.2.0"
peewee = "3.17.1"
proto-plus = "1.23.0"
protobuf = "4.25.2"
py = "1.11.0"
pyarrow = "15.0.1"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pydantic = "2.6.1"
pydantic_core = "2.16.2"
Pygments = "2.17.2"
pylance = "0.10.2"
pyowm = "3.3.0"
PySocks = "1.7.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-iso639 = "2024.2.7"
python-magic = "0.4.27"
pytils = "0.4.1"
pytz = "2024.1"
PyYAML = "6.0.1"
rapidfuzz = "3.6.1"
ratelimiter = "1.2.0.post0"
regex = "2023.12.25"
requests = "2.31.0"
retry = "0.9.2"
rich = "13.7.0"
rsa = "4.9"
sec-api = "1.0.17"
semver = "3.0.2"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
SQLAlchemy = "2.0.27"
tabulate = "0.9.0"
tenacity = "8.2.3"
tiktoken = "0.5.2"
tools = "0.1.9"
tqdm = "4.66.2"
typer = "0.9.0"
typing-inspect = "0.9.0"
typing_extensions = "4.9.0"
tzdata = "2024.1"
unstructured = "0.10.25"
urllib3 = "2.2.0"
webencodings = "0.5.1"
Werkzeug = "3.0.1"
wikipedia = "1.4.0"
wrapt = "1.16.0"
yarl = "1.9.4"
yfinance = "0.2.35"
zipp = "3.17.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"